requirements.txt
dict2xml 1.7.2
python-dotenv